[src] Rename pair_selector to select_pair
